写卡函数
计算密码函数未完成；

+++ LanzhouDemo/app/src/main/java/com/xianfeng/ReadData.java
@@ -12,7 +12,7 @@ public class ReadData {
     String corpno;
     String meterno;
     Integer buycount;
-    Double cardgases;
+    Integer cardgases;
     Double gasfee;
     String buygasdate;
     Integer errorcode;
